backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
requests
python-multipart
faiss-cpu
sentence-transformers
numpy
gunicorn
